🎨[Refactor]: delpoy 모달 컴포넌트 분리 및 UI 변경
